+++ lib/actions/HR/EmployeeData/delete.ts
@@ -1,28 +1,33 @@
-'use server'
+'use server';
 
-import connectToDB from "@/lib/database"
-import EmployeeData from "@/lib/models/HR/EmployeeData.model";
+import { ApiResponse } from '@/interfaces/APIresponses.interface';
+import handleDBConnection from '@/lib/database';
+import EmployeeData from '@/lib/models/HR/EmployeeData.model';
 
-const deleteEmployeeData = async(docId:any) => {
-    try{
-        await connectToDB();
-        const resp = await EmployeeData.findOneAndDelete({
-            _id:docId
-        })
-        return {
-            success:true,
-            status:200,
-            message:'Employee Data Deleted Successfully'
-        }
-    }   
-    catch(err){
-        return{
-            success:false,
-            status:500,
-            message:'Internal Server Error',
-            error:JSON.stringify(err)
-        }
-    }
-}
+const deleteEmployeeData = async (docId: any): Promise<ApiResponse<any>> => {
+  const dbConnection = await handleDBConnection();
+  if (!dbConnection.success) return dbConnection;
+  try {
+    const resp = await EmployeeData.findOneAndDelete({
+      _id: docId,
+    });
+    return {
+      success: true,
+      status: 200,
+      message: 'Employee Data Deleted Successfully',
+      error: null,
+      data: JSON.stringify(resp),
+    };
+  } catch (err) {
+    return {
+      success: false,
+      status: 500,
+      message:
+        'Unexpected error occurred, Failed to update Employee Data, Please Try Later',
+      error: JSON.stringify(err),
+      data: null,
+    };
+  }
+};
 
-export {deleteEmployeeData}
+export { deleteEmployeeData };
