+++ lib/actions/SafetyEmp/weekly/houseKeepingAudit.ts
@@ -1,11 +1,12 @@
-"use server";
+'use server';
 
-import connectToDB from "@/lib/database";
-import HouseKeepingAudit from "@/lib/models/safetyPanel/emp/weekly/houseKeeping.model";
+import handleDBConnection from '@/lib/database';
+import HouseKeepingAudit from '@/lib/models/safetyPanel/emp/weekly/houseKeeping.model';
 
 const genHkAudit = async (dataString: string) => {
   try {
-    await connectToDB();
+    const dbConnection = await handleDBConnection();
+    if (!dbConnection.success) return dbConnection;
     const data = JSON.parse(dataString);
     const docObj = new HouseKeepingAudit({
       ...data,
@@ -14,55 +15,57 @@ const genHkAudit = async (dataString: string) => {
     return {
       success: true,
       status: 200,
-      message: "House Keeping Audit Created Successfully",
+      message: 'House Keeping Audit Created Successfully',
       data: JSON.stringify(resp),
     };
   } catch (err) {
     return {
       success: false,
       status: 500,
-      message: "An Error Occurred",
+      message: 'An Error Occurred',
     };
   }
 };
 
 const deleteHkAudit = async (id: any) => {
   try {
-    await connectToDB();
+    const dbConnection = await handleDBConnection();
+    if (!dbConnection.success) return dbConnection;
     const resp = await HouseKeepingAudit.findByIdAndDelete(id);
     return {
       success: true,
       status: 200,
-      message: "House Keeping Audit Deleted Successfully",
+      message: 'House Keeping Audit Deleted Successfully',
     };
   } catch (err) {
     return {
       success: false,
       status: 500,
-      message: "An Error Occurred",
+      message: 'An Error Occurred',
     };
   }
 };
 
 const fetchAllHkAudit = async () => {
   try {
-    await connectToDB();
+    const dbConnection = await handleDBConnection();
+    if (!dbConnection.success) return dbConnection;
     const resp = await HouseKeepingAudit.find().sort({
-        createdAt:-1
+      createdAt: -1,
     });
     return {
       success: true,
       status: 200,
-      message: "House Keeping Audit Fetched Successfully",
+      message: 'House Keeping Audit Fetched Successfully',
       data: JSON.stringify(resp),
     };
   } catch (err) {
     return {
       success: false,
       status: 500,
-      message: "An Error Occurred",
+      message: 'An Error Occurred',
     };
   }
 };
 
-export {genHkAudit,deleteHkAudit,fetchAllHkAudit}
+export { genHkAudit, deleteHkAudit, fetchAllHkAudit };
