+++ lib/actions/chemicals/deleteChemical.ts
@@ -1,27 +1,32 @@
-'use server'
+'use server';
 
-import connectToDB from "@/lib/database";
-import Chemical from "@/lib/models/safetyPanel/chemicals/chemical.model";
+import { ApiResponse } from '@/interfaces/APIresponses.interface';
+import handleDBConnection from '@/lib/database';
+import Chemical from '@/lib/models/safetyPanel/chemicals/chemical.model';
 
-const deleteChemical = async(chemicalId:any) => {
-    try{
-        await connectToDB();
-        const result = await Chemical.deleteOne({_id:chemicalId});
-        return{
-            success:true,
-            message:'Chemical Deleted',
-            status:200,
-        }
-    }
-    catch(err){
-        console.log(err)
-        return{
-            success:false,
-            status:500,
-            message:'Internal Server Error',
-            err:JSON.stringify(err)
-        }
-    }
-}
+const deleteChemical = async (chemicalId: any): Promise<ApiResponse<any>> => {
+  try {
+    const dbConnection = await handleDBConnection();
+    if (!dbConnection.success) return dbConnection;
+    const result = await Chemical.deleteOne({ _id: chemicalId });
+    return {
+      success: true,
+      message: 'Chemical Deleted',
+      status: 200,
+      data: '',
+      error: null,
+    };
+  } catch (err) {
+    console.log(err);
+    return {
+      success: false,
+      status: 500,
+      message:
+        'Unexpected error occurred, Failed to delete chemical, Please try later',
+      error: JSON.stringify(err),
+      data: null,
+    };
+  }
+};
 
-export default deleteChemical
+export default deleteChemical;
